tests/fixtures/integration/clean_project/requirements.txt
flask==3.0.0
pydantic==2.6.0
PyJWT==2.8.0
httpx==0.27.0
